requirements.txt
astroid==1.6.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
Flask==1.0.2
Flask-OAuthlib==0.9.5
gevent==1.3.2
greenlet==0.4.13
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.1.0
pigar==0.7.2
pylint==1.9.1
python-dotenv==0.8.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
